Move previous rankOutlines tests into its new file
